Moving files for better spliting of tests.
commit-id:5aa218ea
